Extract SettleCallCoordinator from run_loop
Moves `settle`, `wait_for_settlement_transaction`,
`store_execution_started/ended`, and the `SettleError` enum out of
`RunLoop` into a new `run_loop::settle_call_coordinator` submodule.

`RunLoop` holds an `Arc<SettleCallCoordinator>` built once in `new`;
`start_settlement_execution` calls `self.settle_coordinator.settle(...)`
with the same `settle::Request` it always built. Pure refactor — no
behavior change — so the fast-path handler in a later PR can reuse the
submission machinery without ballooning that PR.
